requirements.txt
pandas==2.0.3
requests==2.31.0
tqdm==4.66.1
numpy==1.24.3
Pillow==9.5.0
urllib3==2.0.4
